Refactor the screenshot UI code out of X11
This will now work with any WM implementation we add :)
