+++ internal/ui/embed_wm.go
@@ -1,6 +1,8 @@
 package ui
 
 import (
+	"image"
+
 	"fyne.io/fyne"
 
 	"fyne.io/fynedesk"
@@ -50,6 +52,10 @@ func (e *embededWM) AddStackListener(fynedesk.StackListener) {
 
 func (e *embededWM) Blank() {
 	// no-op, we don't control screen brightness
+}
+
+func (e *embededWM) Capture() image.Image {
+	return nil // would mean accessing the underling OS screen functions...
 }
 
 func (e *embededWM) Close() {
