+++ internal/x11/screenshot_test.go
@@ -1,6 +1,6 @@
 // +build linux openbsd freebsd netbsd
 
-package wm
+package x11
 
 import (
 	"testing"
