+++ internal/x11/win/client.go
@@ -3,6 +3,8 @@
 package win
 
 import (
+	"image"
+
 	"fyne.io/fyne"
 
 	"github.com/BurntSushi/xgb/xproto"
@@ -61,6 +63,10 @@ func NewClient(win xproto.Window, wm x11.XWM) x11.XWin {
 	}
 
 	return c
+}
+
+func (c *client) Capture() image.Image {
+	return x11.CaptureWindow(c.wm.Conn(), c.FrameID())
 }
 
 func (c *client) ChildID() xproto.Window {
