+++ test/desktop.go
@@ -1,6 +1,8 @@
 package test
 
 import (
+	"image"
+
 	"fyne.io/fyne"
 	"fyne.io/fyne/test"
 
@@ -32,6 +34,11 @@ func NewDesktopWithWM(wm fynedesk.WindowManager) *Desktop {
 // AddShortcut is called from modules that wish to register keyboard handlers
 func (*Desktop) AddShortcut(shortcut *fynedesk.Shortcut, handler func()) {
 	// TODO
+}
+
+// Capture the desktop to an image. Our test code cowardly refuses to do this.
+func (*Desktop) Capture() image.Image {
+	return nil // could be implemented if required for testing
 }
 
 // ContentSizePixels returns a default value for how much space maximised apps should use
